+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+### [0.1.0](https://github.com/juliancwirko/buildo-begins/releases/tag/v0.1.0) (2022-06-12)
+- added `buildo-begins issue-esdt` command for issuing new ESDT tokens
+- added `buildo-begins set-special-roles-esdt` command for setting and unsetting special ESDT roles - local mint, and local burn
+- added `buildo-begins mint-burn-esdt` command for minting or burning ESDT token supply (requires special roles from the command above)
+- more operations on ESDT soon. Stay tuned!
+
 ### [0.0.3](https://github.com/juliancwirko/buildo-begins/releases/tag/v0.0.3) (2022-06-04)
 - added `buildo-begins send-esdt` command for sending ESDT tokens
 - added `buildo-begins send-nft` command for sending NFT tokens